Fixed a warning caused by improving the fix for getting rid of the warning that Johnny_cz did originally, the new fix (by me) didn't include a cast needed to fix the warning properly, so this new fix fixes the fix of the fix of the warning :D ... Oh nevermind, it was already fixed by some other fixer, but I put time into writing this commit messages, so I'll just commit it anyway.

+++ q_shared.c
@@ -260,7 +260,7 @@ size_t strlcat(char *dst, const char *src, size_t siz)
 char *strstri(const char *text, const char *find)
 {
 	char *s = (char *)text;
-	size_t findlen = strlen(find);
+	int findlen = strlen(find);
 
 	// Empty substring, return input (like strstr).
 	if (findlen == 0)
